Refactors making requests to own file
